# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AgentRequest {
  /** @minLength 1 */
  sessionId: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  message: string;
}

export interface MemoryState {
  /** @nullable */
  diet: string | null;
  /** @nullable */
  preferredCuisine: string | null;
  /** @nullable */
  dislikedCuisine: string | null;
  /** @nullable */
  spicePreference: string | null;
  /** @nullable */
  budget: number | null;
  /** @nullable */
  location: string | null;
  updatedAt: string;
}

export interface Restaurant {
  id: string;
  name: string;
  city: string;
  cuisine: string[];
  diet: string[];
  rating: number;
  averagePrice: number;
  priceRange: string;
  spiceLevel: string;
  distanceKm: number;
  description: string;
}

export type ActivityEventKind = typeof ActivityEventKind[keyof typeof ActivityEventKind];


export const ActivityEventKind = {
  thought: 'thought',
  tool: 'tool',
  observation: 'observation',
  decision: 'decision',
  success: 'success',
  warning: 'warning',
} as const;

export type ActivityEventStatus = typeof ActivityEventStatus[keyof typeof ActivityEventStatus];


export const ActivityEventStatus = {
  complete: 'complete',
  active: 'active',
  pending: 'pending',
} as const;

export interface ActivityEvent {
  id: string;
  kind: ActivityEventKind;
  label: string;
  detail: string;
  status: ActivityEventStatus;
}

/**
 * @nullable
 */
export type TraceStepArguments = { [key: string]: unknown } | null;

export interface TraceStep {
  step: number;
  title: string;
  type: string;
  detail: string;
  /** @nullable */
  tool: string | null;
  /** @nullable */
  arguments: TraceStepArguments;
  /** @nullable */
  result: string | null;
}

export type RecommendationScoreBreakdown = {[key: string]: number};

export type Recommendation = Restaurant & {
  score: number;
  reason: string;
  scoreBreakdown: RecommendationScoreBreakdown;
};

export type AgentResponseMode = typeof AgentResponseMode[keyof typeof AgentResponseMode];


export const AgentResponseMode = {
  demo: 'demo',
  llm: 'llm',
} as const;

export type AgentResponseStats = {
  searched: number;
  afterCuisine: number;
  afterDiet: number;
  elapsedMs: number;
};

export interface AgentResponse {
  sessionId: string;
  mode: AgentResponseMode;
  reply: string;
  memory: MemoryState;
  activity: ActivityEvent[];
  trace: TraceStep[];
  recommendations: Recommendation[];
  stats: AgentResponseStats;
}

export interface ErrorResponse {
  error: string;
}