# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventKind } from './activityEventKind';
import type { ActivityEventStatus } from './activityEventStatus';

export interface ActivityEvent {
  id: string;
  kind: ActivityEventKind;
  label: string;
  detail: string;
  status: ActivityEventStatus;
}