# lib/api-zod/src/generated/types/agentRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AgentRequest {
  /** @minLength 1 */
  sessionId: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  message: string;
}