# lib/api-zod/src/generated/types/agentResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEvent } from './activityEvent';
import type { AgentResponseMode } from './agentResponseMode';
import type { AgentResponseStats } from './agentResponseStats';
import type { MemoryState } from './memoryState';
import type { Recommendation } from './recommendation';
import type { TraceStep } from './traceStep';

export interface AgentResponse {
  sessionId: string;
  mode: AgentResponseMode;
  reply: string;
  memory: MemoryState;
  activity: ActivityEvent[];
  trace: TraceStep[];
  recommendations: Recommendation[];
  stats: AgentResponseStats;
}