# lib/api-zod/src/generated/types/memoryState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryState {
  /** @nullable */
  diet: string | null;
  /** @nullable */
  preferredCuisine: string | null;
  /** @nullable */
  dislikedCuisine: string | null;
  /** @nullable */
  spicePreference: string | null;
  /** @nullable */
  budget: number | null;
  /** @nullable */
  location: string | null;
  updatedAt: string;
}