# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationScoreBreakdown } from './recommendationScoreBreakdown';
import type { Restaurant } from './restaurant';

export type Recommendation = Restaurant & {
  score: number;
  reason: string;
  scoreBreakdown: RecommendationScoreBreakdown;
};