# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Restaurant {
  id: string;
  name: string;
  city: string;
  cuisine: string[];
  diet: string[];
  rating: number;
  averagePrice: number;
  priceRange: string;
  spiceLevel: string;
  distanceKm: number;
  description: string;
}