# lib/api-zod/src/generated/types/traceStep.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TraceStepArguments } from './traceStepArguments';

export interface TraceStep {
  step: number;
  title: string;
  type: string;
  detail: string;
  /** @nullable */
  tool: string | null;
  /** @nullable */
  arguments: TraceStepArguments;
  /** @nullable */
  result: string | null;
}